Check if old option still exists
git-svn-id: https://plugins.svn.wordpress.org/duplicate-post/trunk@484497 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ duplicate-post-admin.php
@@ -70,24 +70,26 @@ function duplicate_post_plugin_upgrade() {
 		
 		// Get old duplicate_post_copy_user_level option
 		$min_user_level = get_option('duplicate_post_copy_user_level');
-		
-		// Get default roles
-		$default_roles = array(
+
+		if (!empty($min_user_level)){
+			// Get default roles
+			$default_roles = array(
 			1 => 'contributor',
 			2 => 'author',
 			3 => 'editor',
 			8 => 'administrator',
-		);
-			
-		// Cycle all roles and assign capability if its level >= duplicate_post_copy_user_level
-		foreach ($default_roles as $level => $name){
-			$role = get_role($name);
-			if ($role && $min_user_level <= $level)
-			$role->add_cap( 'copy_posts' );
+			);
+				
+			// Cycle all roles and assign capability if its level >= duplicate_post_copy_user_level
+			foreach ($default_roles as $level => $name){
+				$role = get_role($name);
+				if ($role && $min_user_level <= $level)
+				$role->add_cap( 'copy_posts' );
+			}
+
+			// delete old option
+			delete_option('duplicate_post_copy_user_level');
 		}
-		
-		// delete old option
-		delete_option('duplicate_post_copy_user_level');
 		
 		add_option(
 		'duplicate_post_copyexcerpt',
